Group unit tests into separate files by topic
