+++ tests/test_pandas.py
@@ -1,0 +1,58 @@
+import sys
+
+import pytest
+
+from sentinelsat import SentinelAPI
+
+
+@pytest.mark.fast
+@pytest.mark.pandas
+@pytest.mark.geopandas
+def test_missing_dependency_dataframe(monkeypatch):
+    api = SentinelAPI("mock_user", "mock_password")
+
+    with pytest.raises(ImportError):
+        monkeypatch.setitem(sys.modules, "pandas", None)
+        api.to_dataframe({"test": "test"})
+
+    with pytest.raises(ImportError):
+        monkeypatch.setitem(sys.modules, "geopandas", None)
+        api.to_geodataframe({"test": "tst"})
+
+
+@pytest.mark.pandas
+@pytest.mark.scihub
+def test_to_pandas(products):
+    df = SentinelAPI.to_dataframe(products)
+    assert type(df).__name__ == 'DataFrame'
+    assert len(products) == len(df)
+    assert set(products) == set(df.index)
+
+
+@pytest.mark.pandas
+@pytest.mark.fast
+def test_to_pandas_empty():
+    df = SentinelAPI.to_dataframe({})
+    assert type(df).__name__ == 'DataFrame'
+    assert len(df) == 0
+
+
+@pytest.mark.pandas
+@pytest.mark.geopandas
+@pytest.mark.scihub
+def test_to_geopandas(products):
+    gdf = SentinelAPI.to_geodataframe(products)
+    assert type(gdf).__name__ == 'GeoDataFrame'
+    print(gdf.unary_union.area)
+    assert gdf.unary_union.area == pytest.approx(89.6, abs=0.1)
+    assert len(gdf) == len(products)
+    assert gdf.crs == {'init': 'epsg:4326'}
+
+
+@pytest.mark.pandas
+@pytest.mark.geopandas
+@pytest.mark.fast
+def test_to_geopandas_empty():
+    gdf = SentinelAPI.to_geodataframe({})
+    assert type(gdf).__name__ == 'GeoDataFrame'
+    assert len(gdf) == 0
